+++ firmware/Kvantoriada_MEGA/Libs/Robot_libs/Workspace/Workspace.cpp
@@ -5,7 +5,7 @@ Workspace::Workspace()
 
 };
 
-void Workspace::set_cell_size(uint8_t _cell_X_size_, uint8_t _cell_Y_size_, uint8_t _cell_Z_size_)
+void Workspace::set_cell_size(int  _cell_X_size_, int  _cell_Y_size_, int  _cell_Z_size_)
 {
     _cell_X_size = _cell_X_size_;
     _cell_Y_size = _cell_Y_size_;
@@ -13,14 +13,14 @@ void Workspace::set_cell_size(uint8_t _cell_X_size_, uint8_t _cell_Y_size_, uint
 
 };
 
-void Workspace::set_ws_size(uint8_t _ws_X_size_, uint8_t _ws_Y_size_, uint8_t _ws_Z_size_)
+void Workspace::set_ws_size(int  _ws_X_size_, int  _ws_Y_size_, int  _ws_Z_size_)
 {
     _ws_X_size = _ws_X_size_;
     _ws_Y_size = _ws_Y_size_;
     _ws_Z_size = _ws_Z_size_;
 };
 
-uint8_t Workspace::put_Cell_Size(char Cell_Axis)
+int  Workspace::put_Cell_Size(char Cell_Axis)
 {
     switch (Cell_Axis)
     {
@@ -39,7 +39,7 @@ uint8_t Workspace::put_Cell_Size(char Cell_Axis)
     }
 };
 
-uint8_t Workspace::put_WS_Size(char WS_Axis)
+int  Workspace::put_WS_Size(char WS_Axis)
 {
      switch (WS_Axis)
     {
